Add item ID now linked to Trello. Tidy up.
