+++ todo_app/app.py
@@ -1,9 +1,7 @@
-from flask import Flask, render_template, request
+from flask import Flask, render_template, request, url_for
 from werkzeug.utils import redirect
 from todo_app.data.session_items import get_item 
 from todo_app.data.trello_items import complete_item, get_items, add_item
-
-
 from todo_app.flask_config import Config
 
 app = Flask(__name__)
@@ -16,9 +14,9 @@ def index():
 @app.route('/new', methods=['POST'])
 def new_task():
     add_item(request.form.get('title'))
-    return redirect("/")
+    return redirect((url_for("index")))
 
 @app.route('/complete/<id>')
 def complete_task(id):
     complete_item(id)
-    return redirect("/")
+    return redirect(url_for("index"))
