+++ todo_app/data/trello_items.py
@@ -1,51 +1,48 @@
 import os
-from tokenize import String
 import requests
 from flask import session
 from todo_app.data.item import Item
-
-def api_request_get():
-
-    payload = {'key': os.getenv('API_KEY'), 'token': os.getenv('API_TOKEN')}
-    list_ID = os.getenv('OPEN_LIST_ID')
-    get_list_of_items = requests.get( f'https://api.trello.com/1/lists/{list_ID}/cards/', params=payload)
-
-    return get_list_of_items.json()
-    
-def api_request_post(title: String): 
-
-    payload = {'name': title, 'idList': os.getenv('OPEN_LIST_ID'), 'key': os.getenv('API_KEY'), 'token': os.getenv('API_TOKEN')}
-    requests.post( f'https://api.trello.com/1/cards', params=payload)
-
-def api_request_put(card_id): 
-    # Request moves the card into the completed list.  Completed list ID is passed in the payload as idList
-
-    payload = {'key': os.getenv('API_KEY'), 'token': os.getenv('API_TOKEN'), 'idList': os.getenv('CLOSED_LIST_ID')}
-    requests.put( f'https://api.trello.com/1/cards/{card_id}', params=payload)
+from todo_app.data.trello_api_comms import api_request_get, api_request_post, api_request_put
 
 def process_trello_records():
 
-    task_list = set()  #Using a set to avoid duplication
+    task_list = []
+    
     items_dict = api_request_get()
 
     for card in items_dict:
         new_task = Item.from_trello_cards(card)
-        task_list.add(new_task)
+        task_list.append(new_task)
 
     return task_list
 
-def get_items():
+task_list = process_trello_records()
 
-    task_list = process_trello_records()
+def get_items():    
         
     return session.get('items', task_list.copy())
 
 
 def add_item(title):
 
-    api_request_post(title)
+    trello_conf_added = api_request_post(title)
+
+    items = task_list
+
+    id = trello_conf_added['id']    
+
+    new_task = Item.from_trello_cards({'id': id, 'name': title})
+    items.append(new_task)  
     
 
 def complete_item(item_id):
 
+    items = task_list
+
+    for task in items:
+        t = task.id
+        if task.id == item_id:
+            task.status = "Complete"
+    
+    #Updates Trello but not linked to the default items set as yet
     api_request_put(item_id)
